+++ CHANGELOG.md
@@ -17,6 +17,10 @@ ChangeLog
 * #774: Fixes for getting free disk space on Windows.
 * #803: Major changes in the sharing API. If you were using an old sabre/dav
   sharing api, head to the website for more detailed migration notes.
+* #657: Support for optional auth using `{DAV:}unauthorized` and `{DAV:}all`
+  privileges. This allows you to assign a privilege to a resource, allowing
+  non-authenticated users to access it. For instance, this could allow you
+  to create a public read-only collection.
 * #812 #814: ICS/VCF exporter now includes a more useful filename in its
   `Content-Disposition` header. (@Xenopathic).
 * #801: BC break: If you were using the `Href` object before, it's behavior